saving download action on the database
