--- prime-router/assets/csv-download-site/login__inline.html
@@ -1,56 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="utf-8" />
-    <meta http-equiv="X-UA-Compatible" content="IE=Edge" />
-    <meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <meta name="robots" content="none" />
-
-    <title>CDC Prime ReportStream - Login</title>
-    <script src="https://global.oktacdn.com/okta-signin-widget/5.2.1/js/okta-sign-in.min.js"
-      integrity="sha384-pVjuWLMDJj0Bp6nDaWi8nJtFz+GfVP3hxebF4f6z81KDsRK/+DfCoLK/6ZjHWaoh" crossorigin="anonymous"></script>
-    <link href="https://global.oktacdn.com/okta-signin-widget/5.2.1/css/okta-sign-in.min.css" type="text/css" rel="stylesheet"
-      integrity="sha384-NHrrVHQI7Uja/X0/0SMVLSk+nFzWI8AwBi1LqSOu7PogLHanNl0fOq1tTdvgb9C7" crossorigin="anonymous" />
-</head>
-
-<body>
-    <div id="okta-login-container"></div>
-</body>
-
-<script type="text/javascript" th:inline="javascript">
-    var config = {
-        logo: '//logo.clearbit.com/cdc.gov',
-        language: 'en',
-        features: {
-            registration: false, // Enable self-service registration flow
-            rememberMe: false, // Setting to false will remove the checkbox to save username
-            router: true, // Leave this set to true for the API demo
-        },
-        el: "#okta-login-container",
-        baseUrl: 'https://'+ [[${OKTA_baseUrl}]],
-        clientId: [[${OKTA_clientId}]],
-        redirectUri: [[${OKTA_redirect}]],
-        authParams: {
-          issuer: "https://"+[[${OKTA_baseUrl}]]+"/oauth2/default"
-        }
-    };
-
-    function createCookie(name,value,min) {
-        if (min) {
-            var date = new Date();
-            date.setTime(date.getTime()+(min*60*1000));
-            var expires = ";expires="+date.toUTCString();
-        } else var expires = "";
-        document.cookie = name+"="+value+expires+";path=/";
-    }
-
-    var signInWidget= new OktaSignIn(config);
-    signInWidget.showSignInToGetTokens( { scopes: ['openid', 'email', 'profile', 'simple_report'] } )
-        .then( function( tokens ){
-            var token = tokens.accessToken.value;
-            createCookie("jwt", token,5)
-            window.location.replace([[${OKTA_redirect}]]);
-        })
-</script>
-</html>
