--- prime-router/assets/csv-download-site/readme.md
@@ -1,13 +1,0 @@
-# CSV download site
-
-Basic HTML template for downloading daily test result CSVs.
-
-## Setup
-
-This template compiles the US Web Design System (USWDS) with [Dart SASS](https://sass-lang.com/). `/assets/stylesheets/scss/application.scss` imports USWDS and specifies custom styles.
-
-**Watch SCSS and output compressed CSS**
-
-```
-sass --watch --style=compressed  assets/stylesheets/scss/application.scss assets/stylesheets/css/application.css
-```
